third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/coop-navigated-history-popup.https-expected.txt

This is a testharness.js-based test.
FAIL Popup navigating to other origin with COOP: same-origin and back in history assert_equals: name expected "" but got "Popup-navigating-to-other-origin-with-COOP:-same-origin-and-back-in-history"
Harness: the test ran to completion.
